Move open as pdf to action buttons.
Drop duplication in bumblebee overlay, move functionality to action button
rendering. The open as pdf action could theoretically be active wherever we
render the action buttons, we currently choose to do so only in the bumblebee
overlays.
